Added config parsing for yaml file.
