+++ RaceSim/CMakeLists.txt
@@ -9,11 +9,22 @@ FetchContent_Declare(
 )
 FetchContent_MakeAvailable(spdlog)
 
+FetchContent_Declare(
+    yaml-cpp
+    GIT_REPOSITORY https://github.com/jbeder/yaml-cpp.git
+    GIT_TAG 0.8.0
+)
+FetchContent_MakeAvailable(yaml-cpp)
+
 # Project name
 project(RaceSim VERSION 1.0 LANGUAGES CXX)
 
 # Set C++ standard
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD_REQURED 17)
+
+# Add external libraries
+add_subdirectory(libs/date)
+add_subdirectory(libs/pugixml)
 
 # Include src files - GLOB bad practice but good enough for this project
 file(GLOB_RECURSE SOURCES "src/*.cpp")
@@ -21,13 +32,9 @@ message(STATUS "RaceSim src files: ${SOURCES}")
 
 # Add the main executable
 add_executable(RaceSim ${SOURCES})
-target_link_libraries(RaceSim PRIVATE pugixml date spdlog::spdlog)
+target_link_libraries(RaceSim PRIVATE pugixml date spdlog::spdlog yaml-cpp)
 
-# Add external libraries
-add_subdirectory(libs/date)
-add_subdirectory(libs/pugixml)
-
-# Include directories
+# Include directories for the target executable
 target_include_directories(RaceSim PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/include
     ${CMAKE_CURRENT_SOURCE_DIR}/include/config
