+++ RaceSim/data/config/wsc_config.yaml
@@ -1,0 +1,57 @@
+# For Gen 11.5
+/Version: "WSC_2023"
+/Car: # Car characterizationss
+  max_soc: 5.2 # kwh
+  tire_pressure: 5.5 # bar
+  array_area: 3.98 # m^2
+  car_mass: 305 # in kg
+  air_density: 1.17 # in kg/m^3
+  passive_electric_loss: 20 # in w
+  cda: 0.16
+  max_motor_power: 3.5 # in kwh
+  max_car_speed: 120 # in km/h
+  array_power_max: 950 # in W
+  array_efficiency: 0.252 
+  motor_efficiency: 0.9
+  regen_efficiency: 0.8
+  battery_efficiency: 0.95
+
+\LutPaths:
+  base_dir: wsc_2023 # Base directory for all luts gen12_strategy/RaceSim/luts/<base_dir>
+
+  # All following luts are relative to gen12_strategy/RaceSim/luts/<base_dir>
+  base_route_path: /static/baseroute.csv # 3 column csv --> latitude (degrees) | longitude (degrees) | altitude (km)
+  speed_limits_path: /static/longDistanceSpeedLimit.csv # Unused
+  traffic_signals_path: /static/longDistanceTrafficSignals.csv # Unused
+  power_factor_path: /static/powerfactor.csv
+  roll_res_slope_path: /static/rr2.csv
+  roll_res_yint_path: /static/rr1.csv
+  # driving_efficiency: /static/drivingeff.csv # Unused
+  # regen_efficiency: /static/regeneff.csv # Unused
+  # battery_efficiency: /static/batteff.csv # Unused
+  dni_path: /dynamic/dni.csv
+  dhi_path: /dynamic/dhi.csv
+  wind_direction_path: /dynamic/wind_direction_10m.csv
+  wind_speed_path: /dynamic/wind_speed_10m.csv
+
+\Route: # Route conditions
+  control_stops: 2962,5559,9462,11421,14439,16990,20832,23202,25987 # Coordinate indices in the route csv 
+  control_stop_charge_time: 30 # in minutes
+  race_start_time: 08:00:00 # in 24 hour time
+  race_end_time: 17:00:00 # in 24 hour time
+
+\Telemetry: # Current state of the car
+  current_soc: 100 # Percentage out of 100. Starting soc will be calculated as current_soc / 100.0 * max_soc
+  gps_coordinates: -12.46322,130.84618,24.7 # Lat(degrees),Lon(degrees),Alt(metres)
+  current_date_time: 2023-10-22 08:00:00 # YYYY-MM-DD HH:MM:SS LOCAL time. Hour MUST be in 24 hour format
+  utc_adjustment: -9.5 # In hour conversion from local time to utc
+
+\Optimizer:
+  model: Gen 11.5 # Force model of the car to use
+  optimizer: Constant # type of optimization
+  min_speed: 40 # in km/h. Minimum speed to even consider in any solution
+  max_speed: 100 # in km/h. Maximum speed to consider in any solution. Should be lower than max_car_speed
+  num_segments: 1 # Number of segments to split the route into. Note that if optimizer = constant, this is
+                  # 1 by force
+  # Todo: add something for ending soc of each segment and minimum soc (penalized)
+
