+++ RaceSim/include/config/config.h
@@ -9,117 +9,94 @@
 #include <Luts.h>
 #include <set>
 #include <Globals.h>
+#include <custom_time.h>
+#include <unordered_map>
+#include <config_param.h>
 
+/* Define all config parameters */
+#define CONFIG_PARAMETERS\
+	PARAM(max_soc, double, 5.2)\
+	PARAM(tire_pressure, double, 5.5)\
+	PARAM(array_area, double, 3.98)\
+	PARAM(car_mass, double, 305)\
+	PARAM(air_density, double, 1.17)\
+	PARAM(passive_electric_loss, double, 20)\
+	PARAM(cda, double, 0.16) \
+	PARAM(max_motor_power, double, 3.5) \
+	PARAM(max_car_speed, double, 120) \
+	PARAM(array_power_max, double, 950) \
+	PARAM(array_efficiency, double, 0.252) \
+	PARAM(motor_efficiency, double, 0.9) \
+	PARAM(regen_efficiency, double, 0.8) \
+	PARAM(battery_efficiency, double, 0.95) \
+	PARAM(base_dir, std::string, "wsc_2023") \
+	PARAM(base_route_path, std::string, "/static/baseroute.csv") \
+	PARAM(speed_limits_path, std::string, "/static/longDistanceSpeedLimit.csv") \
+	PARAM(traffic_signals_path, std::string, "/static/powerfactor.csv") \
+	PARAM(power_factor_path, std::string, "/static/powerfactor.csv") \
+	PARAM(roll_res_slope_path, std::string, "/static/rr2.csv") \
+	PARAM(roll_res_yint_path, std::string, "/static/rr1.csv") \
+	PARAM(dni_path, std::string, "/dynamic/dni.csv") \
+	PARAM(dhi_path, std::string, "/dynamic/dhi.csv") \
+	PARAM(wind_direction_path, std::string, "/dynamic/wind_direction_10m.csv") \
+	PARAM(wind_speed_path, std::string, "/dynamic/wind_speed_10m.csv") \
+	PARAM(control_stops, std::string, "2962,5559,9462,11421,14439,16990,20832,23202,25987") \
+	PARAM(control_stop_charge_time, double, 30) \
+	PARAM(race_start_time, std::string, "08:00:00") \
+	PARAM(race_end_time, std::string, "17:00:00") \
+	PARAM(current_soc, double, 100) \
+	PARAM(gps_coordinates, Coord, Coord()) \
+	PARAM(current_date_time, std::string, "2023-10-22 08:00:00") \
+	PARAM(utc_adjustment, double, -9.5) \
+	PARAM(model, std::string, "Gen 11.5") \
+	PARAM(optimizer, std::string, "Constant") \
+	PARAM(min_speed, double, 40) \
+	PARAM(max_speed, double, 100) \
+	PARAM(num_segments, double, 1) \
+/*CHANGE GPS_COORDINATES BACK TO COORD DATATYPE*/
 class Config {
 private:
-	/*------------------ All parameters specified by the configuration file ----------------------*/
-	std::set<size_t> control_stop_indices;
-	std::string luts_base_dir;
-	std::string base_route_path;
-	std::string speed_limit_path;
-	std::string traffic_signal_path;
-	std::string dni_path;
-	std::string dhi_path;
-	std::string wind_dir_path;
-	std::string wind_speed_path;
-	std::string power_factor_path;
-	std::string roll_res_slope_path;
-	std::string roll_res_yint_path;
-	std::string drive_eff_path;
-	std::string regen_eff_path;
-	std::string battery_eff_path;
-	int car_type;
+	std::string file_path;
+	static YAML::Node config;
 
-	Coord gps_coordinates;
-	Time* current_date_time;
-	Time* end_date_time;
-	bool stationary_sim;
-	int optimization_type;
-	int num_segments;
-    double max_soc;
-	double tire_pressure;
-	double array_area;
-	double array_eff;
-	double car_mass;
-	double air_density;
-	double passive_elec_loss;
-	int day_end;
-	int day_start;
-	double motor_efficiency;
-	double regen_efficiency; 
-	double battery_efficiency;
-	double cda;
-	double battery_temperature;
-	double starting_soc;
-	double ending_soc;
-	double max_motor_power;
-	double utc_adjustment;
-	double control_stop_charge_time;
-
-	int min_speed;
-	int max_speed;
-	int num_loops;
-	double array_power_max;
-	/*--------------------------------------------------------------------------------------------*/
+	/* Map with all parsed key-value pairs from the yaml config file */
+	static std::unordered_map<std::string, YAML::Node> key_values;
 
 	static Config* instance_ptr;
 
-	/* Load all parameters from xml file. Should only be called from get_instance() */
-	Config(std::string config_file_path);
+	/* Declare all parameters */
+	#define PARAM(name, type, default_value) \
+		Config_Param<type> name;
+		
+	CONFIG_PARAMETERS
 
-	Config() {}
+	#undef PARAM
+
+	/* Initialize all parameters */
+	#define PARAM(name, type, default_value)\
+		name(Config_Param<type>(#name, default_value, key_values)),
+
+	/* Load all parameters from xml file. Should only be called from get_instance() */
+	Config(std::string config_file_path) : CONFIG_PARAMETERS file_path(config_file_path){}
+
+	#undef PARAM
+
+	/* Extract all values from the config */
+	static void get_config_leaf_nodes(YAML::Node node);
 
 public:
+
 	/* Public singleton constructor */
 	static Config* get_instance();
 
 	~Config();
 
 	/* Getters */
-	inline std::set<size_t> get_control_stop_indices() {return control_stop_indices;}
-	inline std::string get_base_route_path() {return base_route_path;}
-	inline std::string get_speed_limit_path() {return speed_limit_path;}
-	inline std::string get_traffic_signal_path() {return traffic_signal_path;}
-	inline std::string get_dni_path() {return dni_path;}
-	inline std::string get_dhi_path() {return dhi_path;}
-	inline std::string get_wind_dir_path() {return wind_dir_path;}
-	inline std::string get_wind_speed_path() {return wind_speed_path;}
-	inline std::string get_power_factor_path() {return power_factor_path;}
-	inline std::string get_roll_res_slope_path() {return roll_res_slope_path;}
-	inline std::string get_roll_res_yint_path() {return roll_res_yint_path;}
-	inline std::string get_drive_eff_path() {return drive_eff_path;}
-	inline std::string get_regen_eff_path() {return regen_eff_path;}
-	inline std::string get_battery_eff_path() {return battery_eff_path;}
-	inline Coord get_gps_coordinates() {return gps_coordinates;}
-	inline Time* get_current_date_time() {return current_date_time;}
-	inline Time* get_end_date_time() {return end_date_time;}
-	inline bool get_stationary_sim() {return stationary_sim;}
-	inline int get_opt_type() {return optimization_type;}
-	inline int get_num_segments() {return num_segments;}
-	inline double get_max_soc() {return max_soc;}
-	inline double get_tire_pressure() {return tire_pressure;}
-	inline double get_array_area() {return array_area;}
-	inline double get_array_eff() {return array_eff;}
-	inline double get_car_mass() {return car_mass;}
-	inline double get_air_density() {return air_density;}
-	inline double get_passive_elec_loss() {return passive_elec_loss;}
-	inline int get_day_end() {return day_end;}
-	inline int get_day_start() {return day_start;}
-	inline double get_motor_efficiency() {return motor_efficiency;}
-	inline double get_regen_efficiency() {return regen_efficiency;}
-	inline double get_battery_efficiency() {return battery_efficiency;}
-	inline double get_cda() {return cda;}
-	inline double get_battery_temperature() {return battery_temperature;}
-	inline double get_starting_soc() {return starting_soc;}
-	inline double get_ending_soc() {return ending_soc;}
-	inline double get_max_motor_power() {return max_motor_power;}
-	inline double get_utc_adjustment() {return utc_adjustment;}
-	inline double get_control_stop_charge_time() {return control_stop_charge_time;}
-	inline int get_min_speed() {return min_speed;}
-	inline int get_max_speed() {return max_speed;}
-	inline int get_num_loops() {return num_loops;}
-	inline double get_array_power_max() {return array_power_max;}
-	inline int get_car_type() {return car_type;}
+	#define PARAM(name, type, default_value) \
+		inline type get_##name() {return name.get_value();}
+
+	CONFIG_PARAMETERS
+	#undef PARAM
 
 	/* No setters since these parameters should never change after initialization */
 };
