+++ RaceSim/libs/pugixml/CMakeLists.txt
@@ -1,7 +1,6 @@
 cmake_minimum_required(VERSION 3.12)
 project(pugixml)
-set(SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/pugixml.cpp)
-message("Pugixml src files: " ${SOURCES})
+message("Pugixml src files: " ${CMAKE_CURRENT_SOURCE_DIR}/src/pugixml.cpp)
 
-add_library(pugixml ${SOURCES})
+add_library(pugixml ${CMAKE_CURRENT_SOURCE_DIR}/src/pugixml.cpp)
 target_include_directories(pugixml PUBLIC include)
