+++ RaceSim/src/model/base_car.cpp
@@ -12,11 +12,11 @@ Car::Car() :
     motor_efficiency(Config::get_instance()->get_motor_efficiency()),
     regen_efficiency(Config::get_instance()->get_regen_efficiency()),
     battery_efficiency(Config::get_instance()->get_battery_efficiency()),
-    passive_electric_loss(Config::get_instance()->get_passive_elec_loss()),
+    passive_electric_loss(Config::get_instance()->get_passive_electric_loss()),
     air_density(Config::get_instance()->get_air_density()),
     array_area(Config::get_instance()->get_array_area()),
     max_soc(Config::get_instance()->get_max_soc()),
     tire_pressure(Config::get_instance()->get_tire_pressure()),
-    array_efficiency(Config::get_instance()->get_array_eff())
+    array_efficiency(Config::get_instance()->get_array_efficiency())
     {
 }
