+++ RaceSim/src/model/car_factory.cpp
@@ -5,8 +5,8 @@
 #include <Globals.h>
 #include <v1_car.h>
 
-Car* Car_Factory::get_car(int car_type) {
-    if (car_type == V1_CAR) {
+Car* Car_Factory::get_car(std::string car_type) {
+    if (car_type == "Gen 11.5") {
         spdlog::info("Using V1 Car Model");
         return (Car*) new V1_Car();
     } else {
