+++ RaceSim/src/simulation/sim.cpp
@@ -133,7 +133,7 @@ void Sim::reset_vars() {
     double max_soc = Config::get_instance()->get_max_soc();
     Coord starting_coord = Config::get_instance()->get_gps_coordinates();
     battery_energy = max_soc;
-    curr_time = *Config::get_instance()->get_current_date_time();
+    //curr_time = *Config::get_instance()->get_current_date_time();
     wind_speed_lut.initialize_caches(starting_coord, curr_time.get_utc_time_point());
     wind_dir_lut.initialize_caches(starting_coord, curr_time.get_utc_time_point());
     dni_lut.initialize_caches(starting_coord, curr_time.get_utc_time_point());
@@ -246,13 +246,13 @@ void Sim::write_result(std::string csv_path) {
 Sim::Sim(Car* model) : 
     car(model),
     wind_speed_lut(Forecast_Lut(Config::get_instance()->get_wind_speed_path())),
-    wind_dir_lut(Forecast_Lut(Config::get_instance()->get_wind_dir_path())),
+    wind_dir_lut(Forecast_Lut(Config::get_instance()->get_wind_direction_path())),
     dni_lut(Forecast_Lut(Config::get_instance()->get_dni_path())),
     dhi_lut(Forecast_Lut(Config::get_instance()->get_dhi_path())),
     control_stop_charge_time(Config::get_instance()->get_control_stop_charge_time()),
-    race_start(Config::get_instance()->get_day_start()),
-    race_end(Config::get_instance()->get_day_end()),
-    starting_coord(Config::get_instance()->get_gps_coordinates()),
-    curr_time(*Config::get_instance()->get_current_date_time())
+    //race_start(Config::get_instance()->get_race_start_time()),
+    //race_end(Config::get_instance()->get_race_end_time()),
+    starting_coord(Config::get_instance()->get_gps_coordinates())
+   // curr_time(*Config::get_instance()->get_current_date_time())
     {
 }
